Additional checks for cross event Change edge index from int to long

+++ uvChecker/src/uvEdge.cpp
@@ -8,7 +8,7 @@ UvEdge::UvEdge()
 {
 }
 
-UvEdge::UvEdge(UvPoint beginPt, UvPoint endPt, int index)
+UvEdge::UvEdge(UvPoint beginPt, UvPoint endPt, long index)
 {
     this->begin = beginPt;
     this->end = endPt;
